do not decode bytes to string by default
It's an incorrect assumption to make that the incoming bytes will always be of type string. What if the client has actually sent an int, or a float or actually an image file that can’t be decoded to string?

The current logic will work fine for clients talking only in terms of strings. But it won’t work otherwise. It’s not a good idea to have this logic in a generic library.
Let’s leave it up to the client to decode the bytes to the appropriate type as they see fit.